requirements.txt
Flask==3.0.0
flask-sock==0.7.0
simple-websocket==1.0.0
cryptography==41.0.7
liboqs-python==0.14.1
